+++ api/src/main/java/org/allaymc/api/eventbus/event/network/ClientConnectEvent.java
@@ -2,7 +2,9 @@ package org.allaymc.api.eventbus.event.network;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 import org.allaymc.api.eventbus.event.CancellableEvent;
+import org.allaymc.api.i18n.MayContainTrKey;
 import org.cloudburstmc.protocol.bedrock.BedrockServerSession;
 
 /**
@@ -12,4 +14,11 @@ import org.cloudburstmc.protocol.bedrock.BedrockServerSession;
 @Getter
 public class ClientConnectEvent extends NetworkEvent implements CancellableEvent {
     protected BedrockServerSession session;
+    /**
+     * The reason that will be shown to the player if the event is cancelled.
+     * If the event is not cancelled, this field will be ignored.
+     */
+    @MayContainTrKey
+    @Setter
+    protected String disconnectReason;
 }